+++ common/src/main/kotlin/dev/sterner/witchery/entity/EntEntity.kt
@@ -14,12 +14,11 @@ import net.minecraft.world.entity.LivingEntity
 import net.minecraft.world.entity.Mob
 import net.minecraft.world.entity.ai.attributes.AttributeSupplier
 import net.minecraft.world.entity.ai.attributes.Attributes
-import net.minecraft.world.entity.ai.goal.LookAtPlayerGoal
-import net.minecraft.world.entity.ai.goal.MeleeAttackGoal
-import net.minecraft.world.entity.ai.goal.RandomLookAroundGoal
-import net.minecraft.world.entity.ai.goal.RandomStrollGoal
+import net.minecraft.world.entity.ai.goal.*
+import net.minecraft.world.entity.ai.goal.target.HurtByTargetGoal
 import net.minecraft.world.entity.ai.goal.target.NearestAttackableTargetGoal
 import net.minecraft.world.entity.monster.Monster
+import net.minecraft.world.entity.npc.Villager
 import net.minecraft.world.entity.player.Player
 import net.minecraft.world.item.enchantment.EnchantmentHelper
 import net.minecraft.world.level.Level
@@ -34,13 +33,23 @@ class EntEntity(level: Level) : Monster(WitcheryEntityTypes.ENT.get(), level) {
         this.setPersistenceRequired()
     }
 
+
     override fun registerGoals() {
+        goalSelector.addGoal(2, MeleeAttackGoal(this, 1.0, false))
+
         goalSelector.addGoal(5, RandomStrollGoal(this, 0.8))
         goalSelector.addGoal(8, RandomLookAroundGoal(this))
         goalSelector.addGoal(3, LookAtPlayerGoal(this, Player::class.java, 3.0f, 1.0f))
         goalSelector.addGoal(4, LookAtPlayerGoal(this, Mob::class.java, 8.0f))
-        goalSelector.addGoal(4, MeleeAttackGoal(this, 1.0, false))
-        targetSelector.addGoal(5, NearestAttackableTargetGoal(this, Player::class.java, true))
+        targetSelector.addGoal(1, HurtByTargetGoal(this))
+        targetSelector.addGoal(
+            2, NearestAttackableTargetGoal(
+                this,
+                Player::class.java, true
+            )
+        )
+        targetSelector.addGoal(3, NearestAttackableTargetGoal(this, Villager::class.java, true))
+
         super.registerGoals()
     }